#1677 Moved master related exceptions to more a more appropriate package.
