+++ platform-pojo-bl/src/main/java/ua/com/fielden/platform/entity/AbstractEntityEditActionProducer.java
@@ -4,6 +4,7 @@ import java.util.function.Supplier;
 
 import ua.com.fielden.platform.entity.factory.EntityFactory;
 import ua.com.fielden.platform.entity.factory.ICompanionObjectFinder;
+import ua.com.fielden.platform.entity_master.exceptions.SimpleMasterException;
 import ua.com.fielden.platform.types.tuples.T2;
 import ua.com.fielden.platform.web.centre.CentreContext;
