+++ platform-pojo-bl/src/main/java/ua/com/fielden/platform/entity/EntityManipulationActionProducer.java
@@ -10,6 +10,7 @@ import com.google.inject.Inject;
 
 import ua.com.fielden.platform.entity.factory.EntityFactory;
 import ua.com.fielden.platform.entity.factory.ICompanionObjectFinder;
+import ua.com.fielden.platform.entity_master.exceptions.SimpleMasterException;
 import ua.com.fielden.platform.error.Result;
 import ua.com.fielden.platform.security.IAuthorisationModel;
 import ua.com.fielden.platform.security.provider.ISecurityTokenProvider;
