+++ platform-pojo-bl/src/main/java/ua/com/fielden/platform/entity/meta/impl/DefaultOpenCompoundMasterActionKeyDefiner.java
@@ -5,9 +5,9 @@ import static org.apache.commons.lang.StringUtils.isEmpty;
 import static ua.com.fielden.platform.reflection.TitlesDescsGetter.getEntityTitleAndDesc;
 
 import ua.com.fielden.platform.entity.AbstractEntity;
-import ua.com.fielden.platform.entity.CompoundMasterException;
 import ua.com.fielden.platform.entity.meta.IAfterChangeEventHandler;
 import ua.com.fielden.platform.entity.meta.MetaProperty;
+import ua.com.fielden.platform.entity_master.exceptions.CompoundMasterException;
 
 /**
  * Default implementation for compound master opener's key property.
